refactor: repository 추상화
