+++ src/main/java/roomescape/repository/JdbcReservationRepository.java
@@ -1,0 +1,91 @@
+package roomescape.repository;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.util.List;
+import javax.sql.DataSource;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.SqlParameterSource;
+import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.stereotype.Repository;
+import roomescape.domain.Reservation;
+import roomescape.domain.ReservationTime;
+
+@Repository
+public class JdbcReservationRepository implements ReservationRepository {
+
+    private final JdbcTemplate jdbcTemplate;
+    private final SimpleJdbcInsert jdbcInsert;
+
+    public JdbcReservationRepository(JdbcTemplate jdbcTemplate, DataSource source) {
+        this.jdbcTemplate = jdbcTemplate;
+        this.jdbcInsert = new SimpleJdbcInsert(source)
+            .withTableName("reservation")
+            .usingGeneratedKeyColumns("id");
+    }
+
+    public Reservation save(Reservation reservation) {
+        SqlParameterSource params = new MapSqlParameterSource()
+            .addValue("name", reservation.getName())
+            .addValue("date", reservation.getDate())
+            .addValue("time_id", reservation.getTimeId());
+        long id = jdbcInsert.executeAndReturnKey(params).longValue();
+        return findById(id);
+    }
+
+    private Reservation findById(Long id) {
+        return jdbcTemplate.queryForObject(
+            """
+                SELECT
+                    r.id as reservation_id,
+                    r.name as reservation_name,
+                    r.date as reservation_date,
+                    t.id as time_id,
+                    t.start_at as time_value
+                FROM reservation as r
+                inner join reservation_time as t
+                on r.time_id = t.id
+                WHERE r.id = ?
+                """,
+            (resultSet, rowNum) -> new Reservation(
+                resultSet.getLong("reservation_id"),
+                resultSet.getString("reservation_name"),
+                LocalDate.parse(resultSet.getString("reservation_date")),
+                new ReservationTime(
+                    resultSet.getLong("time_id"),
+                    LocalTime.parse(resultSet.getString("time_value"))
+                )
+            ), id
+        );
+    }
+
+    public List<Reservation> findAll() {
+        return jdbcTemplate.query(
+            """
+                SELECT
+                    r.id as reservation_id,
+                    r.name,
+                    r.date,
+                    t.id as time_id,
+                    t.start_at as time_value
+                FROM reservation as r
+                inner join reservation_time as t
+                on r.time_id = t.id
+                """,
+            (resultSet, rowNum) -> new Reservation(
+                resultSet.getLong("reservation_id"),
+                resultSet.getString("name"),
+                LocalDate.parse(resultSet.getString("date")),
+                new ReservationTime(
+                    resultSet.getLong("time_id"),
+                    LocalTime.parse(resultSet.getString("time_value"))
+                )
+            )
+        );
+    }
+
+    public void delete(long id) {
+        jdbcTemplate.update("delete from reservation where id = ?", id);
+    }
+}
